+++ s/ok/index.ts
@@ -3,7 +3,7 @@ export * from "./fns/attempt.js"
 export * from "./fns/ok.js"
 export * from "./fns/err.js"
 export * from "./fns/get.js"
-export * from "./fns/need.js"
+export * from "./fns/got.js"
 
 export * from "./types/ok.js"
 export * from "./types/err.js"